requirements.txt
python-telegram-bot==13.2
requests==2.32.3